AIRAVATA-3310: Refactoring
